+++ GameTest/src/Global/Utils.h
@@ -5,8 +5,8 @@
 namespace Utils
 {
 	void GenerateCircleVertices(float radius, const Vector2& center, int steps, std::vector<Vector2>& out_vector);
-	void DrawDot(const Vector2& position, int r, int g, int b);
-	void DrawRectangle(const Vector2& position, int r, int g, int b, const Vector2& size);
+	void DrawDot(const Vector2& position, float r, float g, float b);
+	void DrawRectangle(const Vector2& position, float r, float g, float b, const Vector2& size);
 
 	Vector2& RandomPosition();
 	int RandomInt(int min, int max);
